DEV: Extract a test helper in local-dates
